충돌 해결
